RUE-860: split parser grammar modules
